kafka: move group_metadata_migration in features/migrators
This is just the superficial move of the files and references
to them, without editing them yet.
