+++ src/v/features/migrators/group_metadata.h
