+++ src/v/kafka/CMakeLists.txt
@@ -61,7 +61,6 @@ v_cc_library(
     server/partition_proxy.cc
     server/group_recovery_consumer.cc
     server/group_metadata.cc
-    server/group_metadata_migration.cc
  DEPS
     Seastar::seastar
     v::bytes
